Rewrite to eliminate the database plus other streamline operations.

+++ fake-dedup.py
@@ -20,7 +20,6 @@ import hashlib
 import os
 import time
 import sys
-import shelve
 from stat import *
 
 if __name__ == "__main__":
@@ -37,8 +36,8 @@ if __name__ == "__main__":
     except:
         pass
 
-    cutoff = int(time.time()) - (7 * 24 * 60 * 60)
-    pd = shelve.open(dbdir + "/path2hash", flag='c')
+    #cutoff = int(time.time()) - (7 * 24 * 60 * 60)
+    cutoff = int(time.time()) - 30
     toplev = True
 
     for (dirpath, dirnames, filenames) in os.walk("."):
@@ -47,53 +46,49 @@ if __name__ == "__main__":
         if toplev:
             for d in dirnames:
                 path = os.path.join(dirpath, d)
-                dirstat = os.stat(path, follow_symlinks=False)
+                dirstat = os.stat(path)
                 if dirstat.st_mtime > cutoff:
                     print ("skipping", path, "because it is too new")
                     del dirnames[dirnames.index(d)]
             toplev = False
             continue
+        dirstat = os.stat(dirpath)
         for f in filenames:
             path = os.path.join(dirpath, f)
             filestat = os.stat(path, follow_symlinks=False)
+            if not S_ISREG(filestat.st_mode):
+                continue
+            if filestat.st_size == 0:
+                continue
+            hash = hashlib.md5(open(path, "rb").read()).hexdigest()
+            file = dbdir + '/' + hash
             try:
-                hash = str(pd[path])
+                hashstat = os.stat(file)
+                if filestat.st_ino != hashstat.st_ino:
+                    print ("add", path, "to", hash)
+                    #
+                    # Remove this file and link to the common link point.
+                    # This is where we start reclaiming disk space.
+                    # Pick the earliest of time stamps to set the
+                    # common link point to.
+                    #
+                    os.unlink(path)
+                    os.link(file, path)
+                    if filestat.st_mtime < hashstat.st_mtime:
+                        os.utime(path, (filestat.st_mtime, filestat.st_atime))
             except:
-                hash = None
-            hashstat = None
-            if hash is None:
-                if not S_ISREG(filestat.st_mode):
-                    continue
-                hash = hashlib.md5(open(path, "rb").read()).hexdigest()
-                pd[path] = hash
-                file = dbdir + '/' + hash
-                try:
-                    os.stat(dbdir + '/' + hash, follow_symlinks=False)
-                except:
-                    os.link(path, file, follow_symlinks=False)
-                    os.utime(path, (filestat.st_mtime, filestat.st_atime))
-                    print ("create hash", hash, "for", path)
-                    continue
+                print ("create hash", hash, "for", path)
+                #
+                # First time we've encountered this hash- create the common link point
+                #
+                os.link(path, file)
+        # restore any timestamps to the containing directory
+        os.utime(dirpath, (dirstat.st_mtime, dirstat.st_atime))
 
-            # The hash is in the database and the hash file is present
-            hashstat = os.stat(dbdir + '/' + hash, follow_symlinks=False)
-            if filestat.st_ino != hashstat.st_ino:
-                print ("add", path, "to", hash)
-                os.unlink(path)
-                os.link(dbdir + '/' + hash, path, follow_symlinks=False)
-                os.utime(path, (hashstat.st_mtime, hashstat.st_atime))
-
-    nukelist = []
     for (dirpath, dirnames, filenames) in os.walk(dbdir):
         for f in filenames:
-            if f != "path2hash":
-                path = os.path.join(dirpath, f)
-                stat = os.stat(path, follow_symlinks=False)
-                if stat.st_nlink == 1:
-                    print (path, "has only one link")
-                    nukelist.append(f)
-                    os.unlink(path)
-    for (k, v) in pd.items():
-        if str(v) in nukelist:
-            print("removing", k, "for hash", v)
-            del pd[k]
+            path = os.path.join(dirpath, f)
+            stat = os.stat(path)
+            if stat.st_nlink == 1:
+                print (path, "has only one link so removing")
+                os.unlink(path)
